pyproject.toml
[tool.poetry]
name = "scholarship_app"
version = "0.1.0"
description = "tbd"
authors = ["Brandon Egger <brandonegger64@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poe.tasks]
run = "python -m scholarship_app"
run-server = "poetry run streamlit run scholarship_app/router.py"
lint = "sh scripts/lint.sh"
format = [
  { cmd = "poetry run black scholarship_app" },
  { cmd = "poetry run black tests" },
]
test-e2e = "sh scripts/playwright.sh"
test-unit = "sh scripts/pyunit.sh"
test = "sh scripts/test.sh"
build = [
  { cmd = "poetry export -f requirements.txt --output requirements.txt" },
  { cmd = "poetry run pyoxidizer build" },
  { cmd = "rm ./requirements.txt" },
]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.17.0"
watchdog = "^2.2.1"
pandas = "^1.5.3"
numpy = "^1.24.2"
streamlit-extras = "^0.2.6"
openpyxl = "^3.1.1"
st-pages = "^0.3.2"
streamlit-aggrid = "^0.3.3"
matplotlib = "^3.7.1"
office365-rest-python-client = "^2.4.0"
extra-streamlit-components = "^0.1.56"
fuzzywuzzy = "^0.18.0"
python-dotenv = "^1.0.0"
python-levenshtein = "^0.20.9"
xlwt = "^1.3.0"
streamlit-ace = "^0.1.1"
flet = "^0.5.2"
black = "^23.3.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.16.1"
coverage = "^7.1.0"
asyncio = "3.4.1"
typer = "^0.7.0"
autopep8 = "^2.0.1"
autohooks = "^23.1.0"
autohooks-plugin-pylint = "^22.8.1"
playwright = "^1.32.0"
pytest-playwright = "^0.3.0"
pre-commit = "^3.1.1"
pyoxidizer = "*"

[tool.poetry.scripts]
enopios = "scholarship_app.__main__:app"

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.pylint"]

[tool.autohooks.plugins.pylint]
arguments = ["--rcfile=.pylintrc", "-s", "n"]